third_party/WebKit/LayoutTests/svg/animations/getCurrentTime-pause-unpause-setCurrentTime-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
PASS successfullyParsed is true

TEST COMPLETE
PASS svg.getCurrentTime() is 0.5
PASS svg.getCurrentTime() is 1
PASS svg.getCurrentTime() is 2
PASS svg.getCurrentTime() is 3.5
